Helpers

+++ ci-build.sh
@@ -22,6 +22,12 @@ while [[ $# -gt 0 ]]; do
 	-r | --release)
 		do_release=1
 		;;
+	-i | --install)
+		do_install=1
+		;;
+	-u | --uninstall)
+		do_uninstall=1
+		;;
 	-p | --package)
 		shift
 		case "$1" in
@@ -65,6 +71,8 @@ while [[ $# -gt 0 ]]; do
 		echo "  -t|--tests:         Fills build tests"
 		echo "  -d|--debug:         Build package with debug information"
 		echo "  -r|--release:       Build package without debug information"
+		echo "  -i|--install:       Install after build. May be need root access"
+		echo "  -u|--uninstall:     Uninstall previous build (if build dir still exists)"		
 		echo "  -v|--version:       Build package with custom version"
 		echo "    |--docs:          Generate doxygen documentation"
 		echo "  -h|--help:          Print this message"
@@ -89,7 +97,7 @@ fi
 
 if [[ $do_docs ]]; then
 	if [[ -e docs ]]; then
-		rm -rf docs
+		rm -rf doxy_docs
 	fi
 	doxygen Doxyfile
 fi
@@ -103,20 +111,35 @@ if [[ $do_tests ]]; then
 	TESTS=TRUE
 fi
 
-echo "* Starting CI building script with BUILD_TYPE = $BUILD_TYPE"
+if [[ ${do_uninstall} ]]; then
+  cd build || exit 1
+	sudo make uninstall
+	cd "$src_dir" || exit 1
+	exit 0
+fi
+
+echo "* Start CI building script with BUILD_TYPE = $BUILD_TYPE"
 if [[ -e build ]]; then
 	rm -rf build
 fi
 mkdir build 2> /dev/null && cd build || exit
 cmake .. \
+	-GNinja \
 	-DCMAKE_INSTALL_PREFIX:PATH=/usr/ \
 	-DCMAKE_BUILD_TYPE=$BUILD_TYPE \
 	-DTESTS:BOOL=$TESTS \
-	-DBUILD_VERSION="$BUILD_VERSION"
-cmake --build . -- -j2
+	-DBUILD_VERSION="$BUILD_VERSION" \
+	-DCMAKE_EXPORT_COMPILE_COMMANDS=${TESTS} \
+	-DCMAKE_CXX_CPPCHECK=/usr/bin/cppcheck
+#cmake --build . -- -j2
+ninja
 
 if [[ $do_tests ]]; then
 	ctest -j2
+fi
+
+if [[ ${do_install} ]]; then
+	sudo make install
 fi
 
 if [[ $do_pkg ]]; then
@@ -136,4 +159,7 @@ if [[ $do_pkg ]]; then
 	mkdir -p "$src_dir"/packages/$PACK_TYPE/
 	mv ./*-0."$MINOR"."$BUILD_VERSION"-* "$src_dir"/packages/$PACK_TYPE/ 2> /dev/null
 fi
+
+cd "$src_dir" || exit
+
 exit 0
